Add `gk-roam-insert-link-dwim`
This patch adds a new autoloaded entry point,
`gk-roam-insert-link-dwim`, which prompts the user for a title and
inserts a link to that roam file, creating it if necessary.

As part of this change, most of the guts of `gk-roam-new-at-point` is
refactored into `gk-roam--insert-link-always` so that it can be reused.

A few unused variables are removed.